task delete

+++ src/Api/Endpoints/Tasks/Delete.cs
@@ -13,7 +13,6 @@ public class Delete: Endpoint<Delete.Req, EmptyResponse>
     public class Req
     {
         public int Id { get; set; }
-        public int ParentJobId { get; set; }
     }
 
     public override void Configure()
@@ -25,14 +24,6 @@ public class Delete: Endpoint<Delete.Req, EmptyResponse>
 
     public override async Task HandleAsync(Req req, CancellationToken ct)
     {
-        var job = await JobRepo.GetByIdAsync(req.ParentJobId, ct);
-
-        if (job is null)
-        {
-            await SendNotFoundAsync(ct);
-            return;
-        }
-
         // TODO: check detectors below
         var task = await TaskRepo.FirstOrDefaultAsync(new TaskWithChildrenSpec(req.Id), ct);
         if (task is null)
@@ -40,7 +31,15 @@ public class Delete: Endpoint<Delete.Req, EmptyResponse>
             await SendNotFoundAsync(ct);
             return;
         }
+        
+        var job = await JobRepo.GetByIdAsync(task.Job.Id, ct);
 
+        if (job is null)
+        {
+            await SendNotFoundAsync(ct);
+            return;
+        }
+        
         job.DeleteTask(task);
         await TaskRepo.DeleteAsync(task, ct);
         await JobRepo.SaveChangesAsync(ct);
